+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# REPLACE-GLOBAL-WITH-NEXT-VERSION
+- PT-11292 - Adds possibility to enter separate credentials for sandbox mode in first run wizard
+
 # 1.2.0
 - PT-11233 - Do not show PayPal Express Button on the product detail page, if the product is in clearance
 - PT-11292 - Add possibility to enter separate credentials for sandbox mode